[testing] Refactoring. Add AbstractTestcontainers Logic.
