+++ src/test/java/com/andriiv/amgbackend/AbstractTestcontainers.java
@@ -1,0 +1,38 @@
+package com.andriiv.amgbackend;
+
+import org.flywaydb.core.Flyway;
+import org.junit.jupiter.api.BeforeAll;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+/**
+ * Created by Roman Andriiv (09.08.2023 - 18:42)
+ */
+@Testcontainers
+public abstract class AbstractTestcontainers {
+    @Container
+    protected static final PostgreSQLContainer<?> postgreSQLContainer =
+            new PostgreSQLContainer<>("postgres:latest")
+                    .withDatabaseName("amgbackend-dao-unit-test")
+                    .withUsername("andriiv")
+                    .withPassword("password");
+
+    @BeforeAll
+    static void beforeAll() {
+        Flyway flyway = Flyway.configure().dataSource(
+                postgreSQLContainer.getJdbcUrl(),
+                postgreSQLContainer.getUsername(),
+                postgreSQLContainer.getPassword()).load();
+        flyway.migrate();
+    }
+
+    @DynamicPropertySource
+    private static void registerDataSourceProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", postgreSQLContainer::getJdbcUrl);
+        registry.add("spring.datasource.username", postgreSQLContainer::getUsername);
+        registry.add("spring.datasource.password", postgreSQLContainer::getPassword);
+    }
+}
