--- src/test/java/com/andriiv/amgbackend/AmgBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.andriiv.amgbackend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AmgBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
