+++ src/test/java/com/andriiv/amgbackend/TestcontainersTest.java
@@ -1,49 +1,17 @@
 package com.andriiv.amgbackend;
 
-import org.flywaydb.core.Flyway;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Created by Roman Andriiv (09.08.2023 - 14:56)
  */
-@Testcontainers
-@SpringBootTest
-public class TestcontainersTest {
-    @Container
-    public static final PostgreSQLContainer<?> postgreSQLContainer =
-            new PostgreSQLContainer<>("postgres:latest")
-                    .withDatabaseName("amgbackend-dao-unit-test")
-                    .withUsername("andriiv")
-                    .withPassword("password");
 
-    @DynamicPropertySource
-    private static void registerDataSourceProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgreSQLContainer::getJdbcUrl);
-        registry.add("spring.datasource.username", postgreSQLContainer::getUsername);
-        registry.add("spring.datasource.password", postgreSQLContainer::getPassword);
-    }
-
+public class TestcontainersTest extends AbstractTestcontainers {
     @Test
     void canStartPostgresDB() {
         assertThat(postgreSQLContainer.isRunning()).isTrue();
         assertThat(postgreSQLContainer.isCreated()).isTrue();
-    }
-
-    @Test
-    void canApplyDbMigrationWithFlyway() {
-        Flyway flyway = Flyway.configure().dataSource(
-                postgreSQLContainer.getJdbcUrl(),
-                postgreSQLContainer.getUsername(),
-                postgreSQLContainer.getPassword()).load();
-        flyway.migrate();
-        System.out.println();
     }
 }
